Move utils to its own package
